Add get_absolute_url to Worker Model and Refactor Worker Detail Template
## Worker Model
- Added get_absolute_url method to Worker model
- This method returns the URL for the worker detail page using the reverse function

## Worker Detail Template
- Refactored worker_detail.html template
- Removed unnecessary div for tasks section
- Moved the include of task_list_partial.html to be inline with the tasks heading

Summary:
This commit enhances the Worker model and improves the structure of the worker_detail.html template. The addition of the get_absolute_url method to the Worker model provides a convenient way to generate the URL for the worker detail page, promoting better code organization and reusability. The refactoring of the worker_detail.html template removes an unnecessary div element and aligns the inclusion of the task_list_partial.html with the tasks heading, resulting in a cleaner and more readable template structure.